+++ src/pages/MarketWatchPage/index.tsx
@@ -1,10 +1,14 @@
-import MarketWatch from "../../components/AdvanceTable";
+/**
+ * Internal dependencies.
+ */
+import { AdvanceTable } from "../../components";
+// import useMarketWatchPage from "./useMarketWatchPage";
 
-const marketwatchendpoint = "https://sam.superintegratedapp.com/wp-json/api/stock-data/?selector=stock&selection=all";
+// const marketwatchendpoint =
+//   "https://sam.superintegratedapp.com/wp-json/api/stock-data/?selector=stock&selection=all";
 const MarketWatchPage = () => {
-  return <>
-  <MarketWatch fetchURL={marketwatchendpoint}/>
-  </>;
+  // const { data, columns } = useMarketWatchPage();
+  return <AdvanceTable data={data} columns={columns} />;
 };
 
 export default MarketWatchPage;
